Make "makeDocs.sh" script compatible with regular shells
Currently "npm run spec" command fails, if your default shell does not
support "bashisms". An example of such shell would be "dash", which is
the default shell in Ubuntu and derivative linux distributions.

This happens because even though there is a bash shebang, it is ignored
in the run command in package.json, as it explicitly starts "sh" instead
of just executing. So the shebang does nothing in this cases.

Since the script is simple and doesn't actually need bash features, it
was cleaned up to work with a regular POSIX shell. Additionally fixed
some issues found by shellcheck.